pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Sistema de perguntas e respostas baseado em documentos (RAG) — projeto de aprendizado progressivo."
requires-python = ">=3.11"

# Dependências do projeto
dependencies = [
    "python-dotenv>=1.0.0",        # Carrega variáveis do .env
    "langchain>=1.0.0",             # Framework de orquestração LLM
    "langchain-groq>=1.0.0",        # LLM: Groq (cloud gratuito)
    "langchain-huggingface>=1.0.0",  # Embeddings locais (sentence-transformers)
    "langchain-chroma>=1.0.0",      # Vector store: ChromaDB local
    "langchain-community>=0.4.0",   # Document loaders e ferramentas
]

[project.optional-dependencies]
# Dependências de desenvolvimento (linters, formatters, testes)
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
